requirements.txt
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
easyocr==1.7.2
exceptiongroup==1.2.2
executing==2.1.0
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.9.0
future==1.0.0
idna==3.10
imageio==2.35.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
langdetect==1.0.9
lazy_loader==0.4
Levenshtein==0.26.0
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240927
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-readability-metrics==1.4.5
pyclipper==1.3.0.post5
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
soundfile==0.12.1
SpeechRecognition==3.10.4
stack-data==0.6.3
sympy==1.13.3
tifffile==2024.9.20
tiktoken==0.7.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
opencv-contrib-python==4.10.0.84
fer==22.5.1